backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
flask-jwt-extended
PyJWT
flask-cors
Flask-Limiter

redis
apscheduler
Flask-Mail
flask-wtf

colorlog